Quinn itteration1.1

+++ src/main/java/Scheduler/Arena.java
@@ -2,7 +2,11 @@ package main.java.Scheduler;
 
 import java.math.*;
 
-
+/**
+ * Data Type of arena storing its name, and position
+ * 
+ * @author Quinn Sondermeyer
+ */
 public class Arena {
 	private String name;
 	private float longitude;
@@ -22,12 +26,13 @@ public class Arena {
 
 	/**
 	 * Checks if arena is in a teams radius
+	 * 
 	 * @param latitude2
 	 * @param longitude2
 	 * @param radius
 	 * @return
 	 */
-	public boolean isInrad(float latitude2, float longitude2, double radius) {
+	public boolean isInrad(double latitude2, double longitude2, double radius) {
 		float length = (float) Math.sqrt((Math.pow((longitude-longitude2), 2) + Math.pow((latitude-latitude2), 2)));
 		if (radius < length) {
 			return true;
@@ -36,25 +41,17 @@ public class Arena {
 	}
 	
 	
-	//Getters and Setters
-	
-	/**
-	 * @return the name
+	/*
+	 * Getters and Setters
 	 */
 	public String getName() {
 		return name;
 	}
-
-	/**
-	 * @return the longitude
-	 */
+	
 	public float getLongitude() {
 		return longitude;
 	}
-
-	/**
-	 * @return the latitude
-	 */
+	
 	public float getLatitude() {
 		return latitude;
 	}
